Implemented a Bull Market algorithm change, changed over classes to better fit needs of the implementation

+++ FPTS/src/Equity.java
@@ -72,8 +72,8 @@ public class Equity extends Holding{
 	 * 
 	 * @param newVal - 
 	 */
-	public void setValue(int newVal){
-		int Piama = newVal;
+	public void setValue(double newVal){
+		double Piama = newVal;
 		value = Piama;
 	}
